+++ tests/testthat/test_filterFeatures.R
@@ -20,6 +20,11 @@ test_that("filterFeatures", {
     expect_equal(filter1, filter7)
     expect_identical(lengths(filter1), c(6L, 2L, 1L))
     
+    ## Test filter stored in variable
+    target <- "Mitochondrion"
+    filter8 <- expect_message(filterFeatures(feat1, ~  location == target))
+    expect_equal(filter1, filter8)
+    
     ## Test numerical filters
     filter1 <- expect_message(filterFeatures(feat1, VariableFilter("pval", 0.03, "<=")))
     filter2 <- expect_message(filterFeatures(feat1, ~ pval <= 0.03))